refactor: cut dead post-refactor code and de-pin tests
Remove production indirection left stranded by the drain/connection
refactor, and the tests that pinned the removed seams.

Production:
- collapse no-op default drain-coordinator layer into SpawnDrainLoop
  (delete single_turn_drain.py; plain path uses coordinator=None)
- make SpawnSession.control_actions / raw_terminal_frames_authoritative
  required; drop unreachable `is None` branches in SpawnManager
- delete HarnessConnection.managed_backend / inject_turn / liveness and
  ManagedBackendHandle.parent_death_linked; primary_attach reads
  scope_snapshot directly; resident drain calls resident_backend
  .begin_followup_turn directly

Tests:
- delete dead/vacuous/duplicate coverage and implementation-pinned
  choreography across streaming, state, harness, launch, platform
- split oversized files (watchdog 1343->202+4 modules, reconciliation
  816->511+2); bundle session create/retry and platform scope tests
- consolidate reaper fixtures into integration/state/conftest.py
- make connection test doubles expose resident_backend per base contract

Full suite green (1488 passed), pyright 0 errors, ruff clean.
