+++ tests/integration/harness/test_managed_backend.py
@@ -6,7 +6,6 @@ import asyncio
 import os
 import sys
 from pathlib import Path
-from typing import Any
 
 import psutil
 import pytest
@@ -56,58 +55,9 @@ async def test_launch_managed_backend_records_scope_with_parent_death_outcome(
     runtime_root = resolve_project_runtime_root(tmp_path)
     scopes = read_scopes_from_disk(runtime_root, spawn_id)
 
-    assert handle.parent_death_linked is True
-    assert handle.scope_snapshot.parent_death_linked is True
+    assert handle.scope_handle.snapshot.parent_death_linked is True
     assert [(scope.scope_id, scope.root_pid, scope.parent_death_linked) for scope in scopes] == [
-        ("backend", handle.scope_snapshot.root_pid, True)
-    ]
-    await handle.process.wait()
-
-
-@pytest.mark.asyncio
-async def test_launch_managed_backend_uses_windows_job_parent_death_outcome(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    spawn_id = SpawnId("spawn-win")
-    job_handle: Any = object()
-
-    monkeypatch.setattr(managed_backend, "IS_WINDOWS", True)
-    monkeypatch.setattr(
-        managed_backend,
-        "detached_subprocess_config",
-        lambda: DetachedSubprocessConfig(kwargs={}, parent_death_linked=False),
-    )
-    monkeypatch.setattr(
-        managed_backend,
-        "link_child_lifetime_to_parent",
-        lambda _pid: ParentDeathLink(
-            job_name="meridian-test-job",
-            job_handle=job_handle,
-            parent_death_linked=True,
-        ),
-    )
-
-    handle = await launch_managed_backend(
-        ManagedBackendConfig(
-            spawn_id=spawn_id,
-            harness_id=HarnessId.OPENCODE,
-            command=(sys.executable, "-c", "pass"),
-            cwd=tmp_path,
-            env=os.environ.copy(),
-            control_root=tmp_path,
-        ),
-        stderr=asyncio.subprocess.DEVNULL,
-    )
-
-    assert handle.parent_death_linked is True
-    assert handle.parent_death_link.job_handle is job_handle
-    assert handle.scope_snapshot.parent_death_linked is True
-    assert handle.scope_snapshot.containment == "windows_job"
-    assert handle.scope_snapshot.job_name == "meridian-test-job"
-    scopes = read_scopes_from_disk(resolve_project_runtime_root(tmp_path), spawn_id)
-    assert [(scope.scope_id, scope.root_pid, scope.owner_policy) for scope in scopes] == [
-        ("backend", handle.scope_snapshot.root_pid, "spawn_owned")
+        ("backend", handle.scope_handle.snapshot.root_pid, True)
     ]
     await handle.process.wait()
 
@@ -151,6 +101,6 @@ async def test_launch_managed_backend_records_unknown_birth_sentinel_when_create
     )
 
     scopes = read_scopes_from_disk(resolve_project_runtime_root(tmp_path), spawn_id)
-    assert handle.scope_snapshot.root_created_at_epoch == PROCESS_BIRTH_UNKNOWN_EPOCH
+    assert handle.scope_handle.snapshot.root_created_at_epoch == PROCESS_BIRTH_UNKNOWN_EPOCH
     assert [scope.root_created_at_epoch for scope in scopes] == [PROCESS_BIRTH_UNKNOWN_EPOCH]
     await handle.process.wait()
