+++ tests/integration/launch/test_spawn_manager_hitl.py
@@ -111,6 +111,10 @@ async def test_spawn_manager_codex_hitl_requests_auto_rejected_for_spawned_agent
         def subprocess_pid(self) -> int | None:
             return None
 
+        @property
+        def resident_backend(self) -> None:
+            return None
+
         async def start(self, config: ConnectionConfig, spec: ResolvedLaunchSpec) -> None:
             _ = spec
             self._spawn_id = config.spawn_id
@@ -248,6 +252,10 @@ async def test_spawn_manager_retryable_permission_send_failure_resolves_not_fail
 
         @property
         def subprocess_pid(self) -> int | None:
+            return None
+
+        @property
+        def resident_backend(self) -> None:
             return None
 
         async def start(self, config: ConnectionConfig, spec: ResolvedLaunchSpec) -> None:
