+++ tests/integration/state/test_reaper_reconciliation.py
@@ -12,7 +12,6 @@ test_reaper_managed_primary.py; cancel flows in test_reaper_cancel.py.
 from __future__ import annotations
 
 import time
-from datetime import UTC, datetime
 from pathlib import Path
 
 import pytest
@@ -30,6 +29,9 @@ from tests.integration.state.conftest import (
     _reconcile,
     _record_launch_boundary,
     _write_activity_artifact,
+    _write_report,
+    fake_reaper_liveness,
+    recording_scope_cleanup,
 )
 
 
@@ -77,6 +79,8 @@ def test_reconcile_active_spawn_without_runner_pid_fails_after_startup_grace(
     latest = _get_spawn(runtime_root, spawn_id)
     assert latest.status == "failed"
     assert latest.error == "missing_runner_pid"
+
+
 def test_reconcile_active_spawn_background_without_takeover_evidence_gets_boundary_error(
     tmp_path: Path,
 ) -> None:
@@ -101,6 +105,8 @@ def test_reconcile_active_spawn_background_without_takeover_evidence_gets_bounda
     latest = _get_spawn(runtime_root, spawn_id)
     assert latest.status == "failed"
     assert latest.error == "launch_boundary_no_takeover"
+
+
 def test_reconcile_active_spawn_background_takeover_evidence_keeps_runner_alive_skip(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
@@ -125,10 +131,7 @@ def test_reconcile_active_spawn_background_takeover_evidence_keeps_runner_alive_
         worker_pid=9001,
     )
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: True,
-    )
+    fake_reaper_liveness(monkeypatch, lambda _pid: True)
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -143,10 +146,7 @@ def test_reconcile_active_spawn_returns_unchanged_when_runner_is_alive(
 ) -> None:
     runtime_root, spawn_id = _create_spawn(tmp_path)
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: True,
-    )
+    fake_reaper_liveness(monkeypatch, lambda _pid: True)
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -179,6 +179,8 @@ def test_reconcile_active_spawn_finalizing_stale_heartbeat_marks_orphan_finaliza
     latest = _get_spawn(runtime_root, spawn_id)
     assert latest.status == "failed"
     assert latest.error == "orphan_finalization"
+
+
 def test_reconcile_active_spawn_finalizing_recent_activity_skips(
     tmp_path: Path,
 ) -> None:
@@ -200,15 +202,14 @@ def test_reconcile_active_spawn_finalizing_recent_activity_skips(
     latest = _get_spawn(runtime_root, spawn_id)
     assert latest.status == "finalizing"
     assert latest.error is None
+
+
 def test_reconcile_active_spawn_with_dead_runner_and_no_exit_or_report_fails(
     tmp_path: Path, monkeypatch
 ) -> None:
     runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -232,10 +233,7 @@ def test_reconcile_active_spawn_with_cancel_intent_and_dead_runner_cancels(
         requested_at="2026-06-03T01:00:00Z",
     )
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -247,8 +245,6 @@ def test_reconcile_active_spawn_with_cancel_intent_and_dead_runner_cancels(
 def test_reconcile_active_spawn_with_cancel_intent_keeps_durable_completion(
     tmp_path: Path, monkeypatch
 ) -> None:
-    from tests.integration.state.conftest import _write_report
-
     runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
     spawn_store.record_cancel_intent(
         runtime_root,
@@ -259,10 +255,7 @@ def test_reconcile_active_spawn_with_cancel_intent_keeps_durable_completion(
     )
     _write_report(runtime_root, spawn_id)
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -303,6 +296,8 @@ def test_reconcile_active_spawn_depth_gate_respects_env_matrix(
     if expected_status == "failed":
         assert reconciled.exit_code == 1
         assert latest.exit_code == 1
+
+
 def test_reconcile_active_spawn_dead_runner_recent_activity_still_fails(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
@@ -318,10 +313,7 @@ def test_reconcile_active_spawn_dead_runner_recent_activity_still_fails(
 
     fixed_now = time.time()
     monkeypatch.setattr("meridian.lib.state.reaper.time.time", lambda: fixed_now)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -333,104 +325,12 @@ def test_reconcile_active_spawn_dead_runner_recent_activity_still_fails(
     assert latest.error == "orphan_run"
 
 
-def test_read_projection_does_not_reap_recorded_scope_but_explicit_reconcile_does(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    from meridian.lib.core import process_cleanup
-    from meridian.lib.core.types import SpawnId
-    from meridian.lib.platform.process_scope.base import CleanupResult, ProcessScopeSnapshot
-    from meridian.lib.state.process_scope_projection import record_scope
-
-    runtime_root, spawn_id = _create_spawn(
-        tmp_path,
-        kind="child",
-        harness="opencode",
-        runner_pid=9301,
-        started_at=_OLD_STARTED_AT,
-    )
-    backend_scope = ProcessScopeSnapshot(
-        scope_id="backend",
-        owner_policy="spawn_owned",
-        owner_id=spawn_id,
-        role="harness_backend",
-        containment="pid_tree_fallback",
-        root_pid=9402,
-        root_created_at_epoch=100.0,
-        pgid=None,
-        job_name=None,
-        degraded_reason=None,
-    )
-    record_scope(runtime_root, SpawnId(spawn_id), backend_scope)
-    record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setenv("MERIDIAN_DEPTH", "1")
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
-    terminated_scopes: list[str] = []
-
-    def _terminate_scope(scope, *, grace_seconds: float, reason: str):
-        terminated_scopes.append(f"{scope.scope_id}:{scope.root_pid}:{reason}")
-        return CleanupResult(
-            scope_id=scope.scope_id,
-            root_pid=scope.root_pid,
-            descendant_count=0,
-            reason=reason,
-            grace_seconds=grace_seconds,
-            kill_escalated=False,
-            degraded_fallback=False,
-            skip_reason=None,
-        )
-
-    monkeypatch.setattr(process_cleanup, "terminate_scope_sync", _terminate_scope)
-
-    from meridian.lib.state.reaper import reconcile_spawns
-
-    projected = reconcile_spawns(tmp_path, runtime_root, [record])
-    assert projected[0].status == "failed"
-    assert projected[0].error == "orphan_run"
-    assert _get_spawn(runtime_root, spawn_id).status == "running"
-    assert terminated_scopes == []
-
-    from meridian.lib.ops.spawn import api as spawn_api
-    from meridian.lib.ops.spawn.models import SpawnListInput, SpawnStatsInput, SpawnWaitInput
-
-    listed = spawn_api.spawn_list_sync(
-        SpawnListInput(statuses=(), project_root=tmp_path.as_posix())
-    )
-    assert listed.spawns[0].spawn_id == spawn_id
-    assert listed.spawns[0].status == "failed"
-    stats = spawn_api.spawn_stats_sync(SpawnStatsInput(project_root=tmp_path.as_posix()))
-    assert stats.failed == 1
-    waited = spawn_api.spawn_wait_sync(
-        SpawnWaitInput(
-            spawn_ids=(spawn_id,),
-            timeout=0.01,
-            timeout_explicit=True,
-            poll_interval_secs=0.01,
-            project_root=tmp_path.as_posix(),
-        )
-    )
-    assert waited.spawns[0].status == "failed"
-    assert _get_spawn(runtime_root, spawn_id).status == "running"
-    assert terminated_scopes == []
-
-    monkeypatch.delenv("MERIDIAN_DEPTH", raising=False)
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    assert reconciled.status == "failed"
-    assert reconciled.error == "orphan_run"
-    assert terminated_scopes == ["backend:9402:reaper"]
-
-
 def test_reconcile_active_spawn_dead_runner_reaps_recorded_backend_scope(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from meridian.lib.core import process_cleanup
     from meridian.lib.core.types import SpawnId
-    from meridian.lib.platform.process_scope.base import CleanupResult, ProcessScopeSnapshot
+    from meridian.lib.platform.process_scope.base import ProcessScopeSnapshot
     from meridian.lib.state.process_scope_projection import record_scope
 
     runtime_root, spawn_id = _create_spawn(
@@ -456,26 +356,11 @@ def test_reconcile_active_spawn_dead_runner_reaps_recorded_backend_scope(
     record_scope(runtime_root, SpawnId(spawn_id), backend_scope)
     _write_activity_artifact(runtime_root, spawn_id, "heartbeat", age_secs=5)
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
+    fake_reaper_liveness(monkeypatch, set())
+    terminated_scopes = recording_scope_cleanup(
+        monkeypatch,
+        "meridian.lib.core.process_cleanup.terminate_scope_sync",
     )
-    terminated_scopes: list[str] = []
-
-    def _terminate_scope(scope, *, grace_seconds: float, reason: str):
-        terminated_scopes.append(f"{scope.scope_id}:{scope.root_pid}:{reason}")
-        return CleanupResult(
-            scope_id=scope.scope_id,
-            root_pid=scope.root_pid,
-            descendant_count=0,
-            reason=reason,
-            grace_seconds=grace_seconds,
-            kill_escalated=False,
-            degraded_fallback=False,
-            skip_reason=None,
-        )
-
-    monkeypatch.setattr(process_cleanup, "terminate_scope_sync", _terminate_scope)
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -504,10 +389,7 @@ def test_reconcile_active_spawn_last_attempt_exit_drives_orphan_failure_after_ac
     )
     _write_activity_artifact(runtime_root, spawn_id, "history.jsonl", age_secs=300)
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -542,10 +424,7 @@ def test_reconcile_active_spawn_post_exit_with_live_runner_skips(
         exited_at="1970-01-01T00:00:30Z",
     )
     record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: True,
-    )
+    fake_reaper_liveness(monkeypatch, lambda _pid: True)
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -553,103 +432,53 @@ def test_reconcile_active_spawn_post_exit_with_live_runner_skips(
     latest = _get_spawn(runtime_root, spawn_id)
     assert latest.status == "running"
     assert latest.error is None
-def test_reconcile_active_spawn_finalizes_from_runner_exit_tuple(
+
+
+@pytest.mark.parametrize(
+    ("runner_status", "runner_exit_code", "runner_error", "expected_status", "expected_error"),
+    [
+        ("succeeded", 0, None, "succeeded", None),
+        ("failed", 23, "guardrail_failed", "failed", "guardrail_failed"),
+        ("timed_out", 1, "resident_deadline_expired", "timed_out", "resident_deadline_expired"),
+    ],
+)
+def test_reconcile_active_spawn_finalizes_from_runner_exit_tuple_after_grace(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
+    runner_status: str,
+    runner_exit_code: int,
+    runner_error: str | None,
+    expected_status: str,
+    expected_error: str | None,
 ) -> None:
     runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
     spawn_store.record_runner_exit(
         runtime_root,
         spawn_id,
-        status="succeeded",
-        exit_code=0,
-        error=None,
+        status=runner_status,
+        exit_code=runner_exit_code,
+        error=runner_error,
         exited_at="1970-01-01T00:16:30Z",
     )
     record = _get_spawn(runtime_root, spawn_id)
     monkeypatch.setattr("meridian.lib.state.reaper.time.time", lambda: 1_000.0)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
-    assert reconciled.status == "succeeded"
-    assert reconciled.exit_code == 0
-    assert reconciled.error is None
+    assert reconciled.status == expected_status
+    assert reconciled.exit_code == runner_exit_code
+    assert reconciled.error == expected_error
     latest = _get_spawn(runtime_root, spawn_id)
-    assert latest.status == "succeeded"
-    assert latest.exit_code == 0
-    assert latest.error is None
-def test_reconcile_active_spawn_finalizes_failed_runner_exit_tuple_after_grace(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
-    spawn_store.record_runner_exit(
-        runtime_root,
-        spawn_id,
-        status="failed",
-        exit_code=23,
-        error="guardrail_failed",
-        exited_at="1970-01-01T00:16:30Z",
-    )
-    record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr("meridian.lib.state.reaper.time.time", lambda: 1_000.0)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
-
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    assert reconciled.status == "failed"
-    assert reconciled.exit_code == 23
-    assert reconciled.error == "guardrail_failed"
-    latest = _get_spawn(runtime_root, spawn_id)
-    assert latest.status == "failed"
-    assert latest.exit_code == 23
-    assert latest.error == "guardrail_failed"
-
-
-def test_reconcile_active_spawn_preserves_timed_out_runner_exit_after_grace(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
-    spawn_store.record_runner_exit(
-        runtime_root,
-        spawn_id,
-        status="timed_out",
-        exit_code=1,
-        error="resident_deadline_expired",
-        exited_at="1970-01-01T00:16:30Z",
-    )
-    record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr("meridian.lib.state.reaper.time.time", lambda: 1_000.0)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
-
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    assert reconciled.status == "timed_out"
-    assert reconciled.exit_code == 1
-    assert reconciled.error == "resident_deadline_expired"
-    latest = _get_spawn(runtime_root, spawn_id)
-    assert latest.status == "timed_out"
-    assert latest.exit_code == 1
-    assert latest.error == "resident_deadline_expired"
+    assert latest.status == expected_status
+    assert latest.exit_code == runner_exit_code
+    assert latest.error == expected_error
 
 
 def test_reconcile_active_spawn_durable_report_wins_over_cancelled_runner_exit(
     tmp_path: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    from tests.integration.state.conftest import _write_report
-
     runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
     spawn_store.record_cancel_intent(
         runtime_root,
@@ -669,10 +498,7 @@ def test_reconcile_active_spawn_durable_report_wins_over_cancelled_runner_exit(
     _write_report(runtime_root, spawn_id)
     record = _get_spawn(runtime_root, spawn_id)
     monkeypatch.setattr("meridian.lib.state.reaper.time.time", lambda: 1_000.0)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
+    fake_reaper_liveness(monkeypatch, set())
 
     reconciled = _reconcile(tmp_path, runtime_root, record)
 
@@ -683,134 +509,3 @@ def test_reconcile_active_spawn_durable_report_wins_over_cancelled_runner_exit(
     assert latest.status == "succeeded"
     assert latest.exit_code == 0
     assert latest.error is None
-
-
-def test_reconcile_active_spawn_uses_runner_created_epoch_for_liveness_checks(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    runtime_root, spawn_id = _create_spawn(
-        tmp_path,
-        started_at=_OLD_STARTED_AT,
-        runner_pid=8123,
-    )
-    spawn_store.update_spawn(
-        runtime_root,
-        spawn_id,
-        runner_pid=8123,
-        runner_created_at_epoch=222.25,
-    )
-    observed: dict[str, float | None] = {}
-
-    def _capture_alive(pid: int, created_after_epoch: float | None = None) -> bool:
-        _ = pid
-        observed["created_after_epoch"] = created_after_epoch
-        return False
-
-    monkeypatch.setattr("meridian.lib.state.reaper.is_process_alive", _capture_alive)
-    record = _get_spawn(runtime_root, spawn_id)
-
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    assert observed["created_after_epoch"] == 222.25
-    assert reconciled.status == "failed"
-    assert reconciled.error == "orphan_run"
-
-
-def test_reconcile_active_spawn_falls_back_to_started_epoch_when_runner_created_missing(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    # Force psutil-backed create-time lookup to return None so the test exercises
-    # the started_epoch fallback regardless of whether the fake runner PID happens
-    # to exist on the host (it doesn't on POSIX CI runners, but may on Windows).
-    monkeypatch.setattr(
-        "meridian.lib.state.spawn_store._runner_created_at_epoch_for_pid",
-        lambda _pid: None,
-    )
-    runtime_root, spawn_id = _create_spawn(
-        tmp_path,
-        started_at=_OLD_STARTED_AT,
-        runner_pid=8124,
-    )
-    spawn_store.update_spawn(
-        runtime_root,
-        spawn_id,
-        runner_pid=8124,
-        runner_created_at_epoch=0.0,
-    )
-    spawn_store.update_spawn(runtime_root, spawn_id, runner_pid=8124, runner_created_at_epoch=None)
-    observed: dict[str, float | None] = {}
-
-    def _capture_alive(pid: int, created_after_epoch: float | None = None) -> bool:
-        _ = pid
-        observed["created_after_epoch"] = created_after_epoch
-        return False
-
-    monkeypatch.setattr("meridian.lib.state.reaper.is_process_alive", _capture_alive)
-    record = _get_spawn(runtime_root, spawn_id)
-
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    expected_started_epoch = datetime(2000, 1, 1, tzinfo=UTC).timestamp()
-    assert observed["created_after_epoch"] == expected_started_epoch
-    assert reconciled.status == "failed"
-    assert reconciled.error == "orphan_run"
-
-
-def test_reconcile_active_spawn_finalizing_uses_runner_exit_tuple(
-    tmp_path: Path,
-) -> None:
-    runtime_root, spawn_id = _create_spawn(
-        tmp_path,
-        status="finalizing",
-        started_at=_OLD_STARTED_AT,
-    )
-    spawn_store.record_runner_exit(
-        runtime_root,
-        spawn_id,
-        status="cancelled",
-        exit_code=130,
-        error="cancelled",
-        exited_at="1970-01-01T00:16:30Z",
-    )
-    _write_activity_artifact(
-        runtime_root,
-        spawn_id,
-        "heartbeat",
-        age_secs=300,
-    )
-    record = _get_spawn(runtime_root, spawn_id)
-
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    assert reconciled.status == "cancelled"
-    assert reconciled.exit_code == 130
-    assert reconciled.error == "cancelled"
-
-
-def test_reconcile_active_spawn_runner_exit_grace_skips_until_grace_expires(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    runtime_root, spawn_id = _create_spawn(tmp_path, started_at=_OLD_STARTED_AT)
-    spawn_store.record_runner_exit(
-        runtime_root,
-        spawn_id,
-        status="failed",
-        exit_code=2,
-        error="timeout",
-        exited_at="1970-01-01T00:16:38Z",
-    )
-    record = _get_spawn(runtime_root, spawn_id)
-    monkeypatch.setattr("meridian.lib.state.reaper.time.time", lambda: 1_000.0)
-    monkeypatch.setattr(
-        "meridian.lib.state.reaper.is_process_alive",
-        lambda *_args, **_kwargs: False,
-    )
-
-    reconciled = _reconcile(tmp_path, runtime_root, record)
-
-    assert reconciled == record
-    latest = _get_spawn(runtime_root, spawn_id)
-    assert latest.status == "running"
